+++ Source/MQTTnet/Connecting/MqttClientConnectResult.cs
@@ -30,7 +30,6 @@ public sealed class MqttClientConnectResult
 
     /// <summary>
     ///     Gets a value indicating whether a session was already available or not.
-    ///     MQTTv5 only.
     /// </summary>
     public bool IsSessionPresent { get; init; }
 
@@ -58,7 +57,6 @@ public sealed class MqttClientConnectResult
 
     /// <summary>
     ///     Gets the result code.
-    ///     MQTTv5 only.
     /// </summary>
     public MqttClientConnectResultCode ResultCode { get; init; }
 
@@ -118,4 +116,3 @@ public sealed class MqttClientConnectResult
     ///     MQTTv5 only.
     /// </summary>
     public bool WildcardSubscriptionAvailable { get; init; }
-}